Collection policies

+++ src/app/controllers/collections_controller.rb
@@ -1,22 +1,27 @@
 class CollectionsController < ApplicationController
   before_action :set_collection, only: %i[show update destroy]
-  skip_before_action :authenticate_user!, only: [:show, :index]
-  skip_after_action :verify_authorized, only: [:show, :index]
+  skip_before_action :authenticate_user!, only: %i[show index]
+  skip_after_action :verify_authorized, only: %i[index]
 
   def index
-    @collections = Collection.all
+    # TODO: show @collections from the current user as well
+    @collections = Collection.where(read_state: 'public_read')
   end
 
   def create
     @collection = Collection.new(collection_params)
+    authorize(@collection)
     return @collection if @collection.save
     render json: @collection.errors, status: :unprocessable_entity
   end
 
-  def show; end
+  def show
+    authorize(@collection)
+  end
 
   def update
     @collection = Collection.find_by_id(params[:id])
+    authorize(@collection)
     render json: '', status: :not_found unless @collection
     update_collection_params = collection_params.to_h
     @collection.assign_attributes(update_collection_params)
@@ -26,8 +31,8 @@ class CollectionsController < ApplicationController
 
   def destroy
     @collection = Collection.find_by_id(params[:id])
+    authorize(@collection)
     return render json: '', status: :not_found unless @collection
-    # @TODO: check user delete permissions
     # @TODO: check when collections has associated records, Error:
     # Mysql2::Error: Cannot delete or update a parent row: a foreign key constraint fails ...
     return render json: '', status: :no_content if @collection.destroy
